+++ quick-order/src/main/java/com/quick_order/entity/OrderInfo.java
@@ -31,7 +31,7 @@ public class OrderInfo {
     private Double discountAmount;
     private Double platformChargePerc;
     private Double platformChargeAmount;
-    private Integer outletId;
+    private Long outletId;
     private Integer orderQty;
     private String preparationNotes;
     private Integer status;
